platform: fix type in formatter string

+++ engine/platform/posix/sys_posix.c
@@ -56,7 +56,7 @@ void Posix_Daemonize( void )
 		if( daemon > 0 )
 		{
 			// parent
-			Con_Reportf( "Child pid: %i\n", daemon );
+			Con_Reportf( "Child pid: %ld\n", (long)daemon );
 			exit( 0 );
 		}
 		else
